requirements.txt
flask
python-dotenv
flask_cors
jsonschema
flask_jwt_extended
cryptography
tensorflow
keras
numpy
pandas